chore: renomear tabela expense_payment para reconciliation e alterar tipo do campo amount para integer

+++ database/migrations/2026_03_16_160941_create_expenses_table.php
@@ -40,7 +40,7 @@ return new class extends Migration
                 ->cascadeOnDelete();
 
             $table->string('reference')->nullable();
-            $table->decimal('amount', 12, 2)->default(0);
+            $table->integer('amount')->default(0);
             $table->date('due_at')->nullable()->index(); // vencimento
             $table->date('competence_date')->nullable()->index(); // vencimento
             $table->string('description')->nullable();
